+++ src/account/validators/password_complex.py
@@ -29,34 +29,3 @@ class ComplexPasswordValidator:
             "Your password must contain at least one uppercase letter, one lowercase letter, "
             "one number, and one symbol."
         )
-
-class NoSpacePasswordValidator:
-    """
-    Validates that a password does not contains a space.
-    """
-
-    def validate(self, password, user=None):
-        if ' ' in password:
-            raise ValidationError(_('Password cannot contain spaces.'), code='password_invalid_character')
-
-    def get_help_text(self):
-        return _(
-            "Your password can't contain a space."
-        )
-
-class ValidCharsetPasswordValidator:
-    """
-    Validates that a password only contains valid characters.
-    """
-
-    def __init__(self, allowed_charset="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()-_+=}{][|:;'\"><,.?/\\"):
-        self.allowed_charset = allowed_charset
-
-    def validate(self, password, user=None):
-        if not all(char in self.allowed_charset for char in password):
-            raise ValidationError(_('Password must contain valid characters.'), code='password_invalid_character')
-
-    def get_help_text(self):
-        return _(
-            "Your password must only contain valid ASCII characters."
-        )
